Revert changes to CroppedRegionSet margins
Though not mentioned in the initial commit messages, this reverts the
fix done for issue #17 where the bottom of a cropped map was cut off.

The issue seen in #17 is nothing to do with a cropped map; it's easily
reproducible (in a much more severe form) with a world that only
contains chunk (0,0) without even considering a crop.

This reverts commit 3a1bb47aa0023549d96f23744a25105caf1df3c3.
This reverts commit 8be3b239b0934c857bf087f3db80326edbcd732a.
This reverts commit 291a6608beaafe7c6b61692ccdcc759911c036b7.

Fixes #90

Change-Id: I02698911647cf829096f20aecbfa01bb8fd375ca